[GStreamer][WebRTC] increase size of apps max-bytes for video tracks https://bugs.webkit.org/show_bug.cgi?id=305576
Reviewed by Xabier Rodriguez-Calvar.

The default value for max-bytes in 200000 is not enough and leads to
frame drops during pipeline buildup.

See: https://github.com/WebPlatformForEmbedded/WPEWebKit/pull/1595

This patch increases the buffering size of the appsrc element used for
video tracks in GStreamerMediaStreamSource so that it can handle enough
buffering to allow for the rest of the pipeline to be built before
buffers start to be lost. The right types are used for the parameters,
according to GObject best practices[1].

Aa an opportunistic cleanup task, types for max-buffer/max-bytes
arguments in GStreamerVideoDecoderFactory are also enforced as guint64.
This seems to be the recommended way to do it according to the GObject
documentation[1],

[1] https://docs.gtk.org/gobject/method.Object.set.html#description

Original author: Eugene Mutavchi <Ievgen_Mutavchi@comcast.com>

* Source/WebCore/platform/mediastream/gstreamer/GStreamerMediaStreamSource.cpp: Increase buffering size for video tracks and use proper types for the parameters.
* Source/WebCore/platform/mediastream/libwebrtc/gstreamer/GStreamerVideoDecoderFactory.cpp: Cleanup by properly typing guint64 parameters.

Canonical link: https://commits.webkit.org/305705@main

+++ Source/WebCore/platform/mediastream/libwebrtc/gstreamer/GStreamerVideoDecoderFactory.cpp
@@ -92,7 +92,7 @@ public:
     bool Configure(const webrtc::VideoDecoder::Settings& codecSettings) override
     {
         m_src = makeElement("appsrc"_s);
-        g_object_set(m_src, "is-live", TRUE, "do-timestamp", TRUE, "max-buffers", 2, "max-bytes", 0, nullptr);
+        g_object_set(m_src, "is-live", TRUE, "do-timestamp", TRUE, "max-buffers", static_cast<guint64>(2), "max-bytes", static_cast<guint64>(0), nullptr);
 
         GRefPtr<GstCaps> caps = nullptr;
         auto capsfilter = CreateFilter();
